Sea scene added
Sea v1 using Game2DWaterKit plugin.
I don't know whether the scale is correct, nor how to test whether it's correct? The scene is designed at full zoom-out view, so should be expanded when zoomed into the cruiser. Also note @Felix: There is a water material property called Surface Color, which I've set to white, but this needs to be set to a different colour depending on the colour of the sky. At the bottom of each sky, there is a constant colour, to which the surface color needs to be set dynamically depending on the sky.